formatting and reorg some parts

+++ rpi.sh
@@ -1,4 +1,3 @@
- 
 echo
 echo "************************************"
 echo "****** Raspberry Pi setup... *******"
@@ -10,7 +9,6 @@ echo
 ##############################################################
 echo
 echo "**** Boot config settings... *****"
-echo
 
 if ! grep -q "max_usb_current=1" "/boot/config.txt"; then
     echo "max_usb_current=1" >>/boot/config.txt
@@ -32,29 +30,30 @@ fi
 #echo "sdram_freq=450" >>boot/config.txt
 #echo "core_freq=450" >>boot/config.txt
 
+echo "...done"
+
 ##############################################################
 ##  Disable serial console
 ##############################################################
 echo
 echo "**** Disable serial console... *****"
-echo
 
 sed -i /boot/cmdline.txt -e "s/console=ttyAMA0,[0-9]\+ //"
+
+echo "...done"
 
 ##############################################################
 ##  Edimax wifi dongle check
 ##############################################################
 echo
 echo "**** Edimax wifi dongle check... *****"
-echo
-### Check if RPi2 with Edimax Wifi dongle
+
 if [ "$EW7811Un" != '' ]; then
-    echo
-    echo "**** Edimax wifi dongle found, building hostapd... *****"
-    echo
+    echo "edimax wifi dongle found, copying the hostapd binaries... *****"
 
     cd $SCRIPTDIR
     rm -f /usr/sbin/hostapd
+    rm -f /usr/sbin/hostapd_cli
 
     #echo "hostapd edimax source"
     #### http://www.edimax.com/images/Image/Driver_Utility/Wireless/NIC/EW-7811Un/EW-7811Un_Linux_driver_v1.0.0.5.zip
@@ -70,9 +69,19 @@ if [ "$EW7811Un" != '' ]; then
         exit 0
     fi
 
+    gunzip -k hostapd_cli.gz
+    if [ ! -f ./hostapd_cli ]; then
+        echo "ERROR - hostapd_cli doesn't exist, exiting..."
+        exit 0
+    fi
+
     # install the binary
     mv ./hostapd /usr/sbin/hostapd
     chmod +x /usr/sbin/hostapd
+
+    # install the binary
+    mv ./hostapd_cli /usr/sbin/hostapd_cli
+    chmod +x /usr/sbin/hostapd_cli
 
     #cd $SCRIPTDIR
     #rm -rf wpa_supplicant_hostapd/
@@ -80,7 +89,11 @@ if [ "$EW7811Un" != '' ]; then
     if ! grep -q "options 8192cu rtw_power_mgnt=0 rtw_enusbss=0" "/etc/modprobe.d/8192cu.conf"; then
         echo "options 8192cu rtw_power_mgnt=0 rtw_enusbss=0" >>/etc/modprobe.d/8192cu.conf
     fi
+else
+    echo "edimax wifi dongle not found, nothing to do... *****"
 fi
+
+echo "...done"
 
 ##############################################################
 ##  I2C setup
@@ -97,12 +110,13 @@ if ! grep -q "i2c-dev" "/etc/modules"; then
     echo "i2c-dev" >>/etc/modules
 fi
 
+echo "...done"
+
 ##############################################################
 ##  Sysctl tweaks
 ##############################################################
 echo
 echo "**** Sysctl tweaks... *****"
-echo
 
 if grep -q "net.core.rmem_max" "/etc/sysctl.conf"; then
     line=$(grep -n 'net.core.rmem_max' /etc/sysctl.conf | awk -F':' '{print $1}')d
@@ -129,3 +143,4 @@ echo "net.core.rmem_default = 167772160" >>/etc/sysctl.conf
 echo "net.core.wmem_max = 167772160" >>/etc/sysctl.conf
 echo "net.core.wmem_default = 167772160" >>/etc/sysctl.conf
 
+echo "...done"
