+++ stratux-setup.sh
@@ -5,11 +5,27 @@ if [ $(whoami) != 'root' ]; then
     exit 0
 fi
 
+BLACK=$(tput setaf 0)
+RED=$(tput setaf 1)
+GREEN=$(tput setaf 2)
+YELLOW=$(tput setaf 3)
+LIME_YELLOW=$(tput setaf 190)
+POWDER_BLUE=$(tput setaf 153)
+BLUE=$(tput setaf 4)
+MAGENTA=$(tput setaf 5)
+CYAN=$(tput setaf 6)
+WHITE=$(tput setaf 7)
+BRIGHT=$(tput bold)
+NORMAL=$(tput sgr0)
+BLINK=$(tput blink)
+REVERSE=$(tput smso)
+UNDERLINE=$(tput smul)
+
 SCRIPTDIR="`pwd`"
 
 cd /root
 
-#set -e 
+#set -e
 
 #outfile=setuplog
 #rm -f $outfile
@@ -35,7 +51,7 @@ echo
 echo "************************************"
 echo "**** Stratux Setup Starting... *****"
 echo "************************************"
-echo 
+echo
 
 ntpd -q -g
 
@@ -44,7 +60,6 @@ ntpd -q -g
 ##############################################################
 echo
 echo "**** Installing dependencies... *****"
-echo
 
 apt-get install -y git
 git config --global http.sslVerify false
@@ -55,40 +70,51 @@ apt-get install -y wget
 apt-get install -y screen
 apt-get install -y isc-dhcp-server
 apt-get install -y tcpdump
-apt-get install -y cmake 
-apt-get install -y libusb-1.0-0.dev 
+apt-get install -y cmake
+apt-get install -y libusb-1.0-0.dev
 apt-get install -y build-essential
 apt-get install -y mercurial
-apt-get install -y autoconf 
-apt-get install -y fftw3 
+apt-get install -y autoconf
+apt-get install -y fftw3
 apt-get install -y fftw3-dev
 apt-get install -y libtool
 apt-get install -y automake
 apt-get install -y hostapd
 apt-get install -y rfkill
 
+echo "...done"
+
 ##############################################################
-##  Platform and hardware specific items
+##  Hardware checkout
 ##############################################################
+echo
+echo "**** Hardware checkout... *****"
+
 REVISION="$(cat /proc/cpuinfo | grep Revision | cut -d ':' -f 2 | xargs)"
 if [ "$REVISION" == "$RPI2BxREV" ] || [ "$REVISION" == "$RPI2ByREV" ]  || [ "$REVISION" == "$RPI3BxREV" ]; then
     echo
     echo "**** Raspberry Pi detected... *****"
-    echo
+
     source $SCRIPTDIR/rpi.sh
 elif [ "$REVISION" == "$ODROIDC2" ]; then
+    echo
+    echo "**** Odroid-C2 detected... *****"
+
     source $SCRIPTDIR/odroid.sh
 else
+    echo
     echo "**** Unable to identify the board using /proc/cpuinfo, exiting *****"
+
     exit 0
 fi
+
+echo "...done"
 
 ##############################################################
 ##  SSH steup and config
 ##############################################################
 echo
 echo "**** SSH setup and config... *****"
-echo
 
 if [ ! -d "$DIRECTORY" ]; then
     mkdir -p /etc/ssh/authorized_keys
@@ -103,13 +129,13 @@ cp -n /etc/ssh/sshd_config{,.bak}
 cp -f $SCRIPTDIR/files/sshd_config /etc/ssh/sshd_config
 rm -f /usr/share/dbus-1/system-services/fi.epitest.hostap.WPASupplicant.service
 
+echo "...done"
 
 ##############################################################
 ##  Hardware blacklisting
 ##############################################################
 echo
 echo "**** Hardware blacklisting... *****"
-echo
 
 if ! grep -q "blacklist dvb_usb_rtl28xxu" "/etc/modprobe.d/rtl-sdr-blacklist.conf"; then
     echo blacklist dvb_usb_rtl28xxu >>/etc/modprobe.d/rtl-sdr-blacklist.conf
@@ -123,12 +149,13 @@ if ! grep -q "blacklist rtl2832" "/etc/modprobe.d/rtl-sdr-blacklist.conf"; then
     echo blacklist rtl2832 >>/etc/modprobe.d/rtl-sdr-blacklist.conf
 fi
 
+echo "...done"
+
 ##############################################################
 ##  Go environment setup
 ##############################################################
 echo
 echo "**** Go environment setup... *****"
-echo
 
 # if any of the following environment variables are set in .bashrc delete them
 if grep -q "export GOROOT_BOOTSTRAP=" "/root/.bashrc"; then
@@ -152,7 +179,7 @@ if grep -q "export PATH=" "/root/.bashrc"; then
 fi
 
 # only add new paths
-XPATH=
+XPATH='$PATH'
 if [[ ! "$PATH" =~ "/root/go/bin" ]]; then
     XPATH+=:/root/go/bin
 fi
@@ -164,16 +191,17 @@ fi
 echo export GOROOT_BOOTSTRAP=/root/gobootstrap >>/root/.bashrc
 echo export GOPATH=/root/gopath >>/root/.bashrc
 echo export GOROOT=/root/go >>/root/.bashrc
-echo export PATH=\$PATH$XPATH >>/root/.bashrc
+echo export PATH=$XPATH >>/root/.bashrc
 
 source /root/.bashrc
+
+echo "...done"
 
 ##############################################################
 ##  Go bootstrap compiler installtion
 ##############################################################
 echo
 echo "**** Go bootstrap compiler installtion... *****"
-echo
 
 cd /root
 
@@ -182,26 +210,31 @@ rm -rf gobootstrap/
 
 wget https://storage.googleapis.com/golang/go1.6.linux-armv6l.tar.gz
 tar -zxvf go1.6.linux-armv6l.tar.gz
-mv go gobootstrap
-rm -f go1.6.linux-armv6l.tar.gz
+if [ ! -d /root/go ]; then
+    echo "Error - go folder doesn't exist, exiting..."
+    exit 0
+fi
 
+rm -f go1.6.linux-armv6l.tar.gz
 rm -rf /root/gopath
 mkdir -p /root/gopath
+
+echo "...done"
 
 ##############################################################
 ##  Go host compiler build
 ##############################################################
 echo
 echo "**** Go host compiler build... *****"
-echo
 
 cd /root
 
 if [ "$REVISION" == "$RPI2BxREV" ] || [ "$REVISION" == "$RPI2ByREV" ]; then
-    mv gobootstrap go
-else
     #### For RPi-2/3, is there any disadvantage to using the armv6l compiler?
     #### to compiling from source?
+    echo "...not necessary, done"
+else
+    mv go gobootstrap
     wget https://storage.googleapis.com/golang/go1.6.src.tar.gz
     tar -zxvf go1.6.src.tar.gz
     rm go1.6.src*
@@ -212,6 +245,8 @@ else
 
     cd /root
     rm -rf gobootstrap/
+
+    echo "...done"
 fi
 
 ##############################################################
@@ -219,7 +254,6 @@ fi
 ##############################################################
 echo
 echo "**** RTL-SDR library build... *****"
-echo
 
 cd /root
 
@@ -233,12 +267,13 @@ make
 make install
 ldconfig
 
+echo "...done"
+
 ##############################################################
 ##  Stratux build and installation
 ##############################################################
 echo
 echo "**** Stratux build and installation... *****"
-echo
 
 cd /root
 
@@ -250,13 +285,13 @@ git checkout v0.8r1
 make all
 make install
 
+echo "...done"
 
 ##############################################################
-##  Kalibrate build and installation build
+##  Kalibrate build and installation
 ##############################################################
 echo
-echo "**** Kalibrate build and installation build... *****"
-echo
+echo "**** Kalibrate build and installation... *****"
 
 cd /root
 
@@ -268,12 +303,13 @@ cd kalibrate-rtl
 make
 make install
 
+echo "...done"
+
 ##############################################################
 ##  System tweaks
 ##############################################################
 echo
 echo "**** System tweaks... *****"
-echo
 
 ##### disable serial console
 if [ -f /etc/inittab ]; then
@@ -290,12 +326,13 @@ if [ -f /usr/sbin/policy-rc.d ]; then
     rm /usr/sbin/policy-rc.d
 fi
 
+echo "...done"
+
 ##############################################################
 ##  WiFi Access Point setup
 ##############################################################
 echo
 echo "**** WiFi Access Point setup... *****"
-echo
 
 source /$SCRIPTDIR/wifi-ap.sh
 
@@ -304,7 +341,7 @@ if which ntp >/dev/null; then
     update-rc.d ntp disable
 fi
 
-#### 
+####
 update-rc.d stratux enable
 
 echo
